ESP32 FINAL FILES
This commit contains all of the expected use files for the ESP32 project. Most are backwards compatible and can be easily updated to match the Arduino system, but the initial design for the platformio repo and it's esp32 components will be overwritten to adjust for the arduino usage, so this commit will act as proof of what I accomplished before needing to pivot my approach.