Fix: Updated vars for MD
Updated variables for the motor driver pins to reflect the correct locations on the Arduino board.
Updated the platformio.ini file to work with the proper arduino file installs rather than esp32 systems.

The current motortesting file is confirmed to be functional with arduino systems in getting the motors to move properly.