Speed Adjustments
Updated the motor pin for ENA to send speed, as pin8 is not PWM enabled, though the correct possible pins are indicated on the ENA pin setter.

Notes on motor speed:
Maximum motor speed is untested, but stated to be 255. Minimum is 128 as the motor's power requirements eventually stop it from moving entirely at speed=128. However, the speed falloff seems to be exponentially decreasing from 255 to 128, with the currently set speed of 150 being still extremely fast. It is currently unclear whether the speed setting adjusts voltage, current, or both; hopefully it should allow for lowered speeds while maintaining high torque, but I'm not too optimistic about that.

The current motor speeds at 150 have also been tested with the main tank tread - while it was functional, the initial 200 speed input ripped off a few of the pins after accidentally touching a part of the table. Shifting down to 150 made it more stable but it was still moving fast enough to where I'm not sure it would be able to move the tank at an appropriate speed without rubber attachments to the treads. I will need to find some form of rubber to cut and attach to the main tread body to facilitate better movement of the tank.

Also, at the current pace of progress, I am unoptimistic that I'd be able to finish the laser pointer system; I will continue to try but there are many factors against it, the lid will be extremely difficult to design with the laser pointer attached, and the overall need for it is low, as testing has shown that nearby cats were entirely uninterested in the laser, and much more interested in bare feet, fingers, and tangible moving objects.