Turret Code Design
Designed the initial turret systems, testing the motor controls of the 6v 1.6kg mini motor running from the arduino normally, as well as the laser pointer system. I also designed a small function to randomise time with the laser on and off, but favouring previous settings to add longer time delay and more natural switching.