Designing Triangular Randomisation
This s an implementation of my idea for a randomised value generation system that prioritises closer values to the previous output. E.g. a previous angle of 50 would randomly output anything from -361 to 361, but with heavier weighting towards values nearby to 50.

I've attempted to integrate this system into the motor rotation system, but it currently has multiple unresolved issues that need adjustment.