Updates to main logic for Arduino code
Added the primary manual movement controls for interpretation from the bluetooth module. Each is localised to a function that is called based on the input from the PC. This is, naturally, only accessible via the manual mode, though I am planning to reuse the functions for the automated movement system since they're so convenient.

Some functions remain as placeholders for the time being, but will be updated accordingly soon enough.

Also, I originally planned to use case switching for the receiver checker, but apparently case switching only works for integers and number values in arduino systems, so I'm stuck with a massive if-else loop unless I can find some way to optimise.